Add new model version to remove property
The wasScheduled boolean property of the NotificationMO entitiy
was sometimes causing crashes, maybe because it was being
modified from a different queue, not sure. Since the attribute
isn't needed any longer, (the user notifications sometimes fail,
and there'll be a new manager to re-schedule them every time
the app becomes active) there's no need to keep it.